change operatornode generic defaults
